Refactor existing inks table
Split in smaller components and add unit tests
