+++ app/javascript/src/collected_inks/table/__tests__/utils.spec.js
@@ -1,0 +1,149 @@
+// @ts-check
+import { booleanSort, colorSort, fuzzyMatch } from "../utils";
+
+describe("colorSort", () => {
+  it("sorts the given color column in ascending hexadecimal value order", () => {
+    const input = [
+      { values: { color: "#fff" } },
+      { values: { color: "#000" } },
+      { values: { color: "#ccc" } }
+    ];
+
+    const expected = [
+      { values: { color: "#000" } },
+      { values: { color: "#ccc" } },
+      { values: { color: "#fff" } }
+    ];
+
+    const result = input.sort((a, b) => colorSort(a, b, "color"));
+
+    expect(result).toStrictEqual(expected);
+  });
+});
+
+describe("booleanSort", () => {
+  it("sorts the given bool column in order of false to true", () => {
+    const input = [
+      { values: { swabbed: true } },
+      { values: { swabbed: false } }
+    ];
+
+    const expected = [
+      { values: { swabbed: false } },
+      { values: { swabbed: true } }
+    ];
+
+    const result = input.sort((a, b) => booleanSort(a, b, "swabbed"));
+
+    expect(result).toStrictEqual(expected);
+  });
+});
+
+describe("fuzzyMatch", () => {
+  const input = [
+    {
+      values: {
+        brand_name: "Pilot",
+        line_name: "Iroshizuku",
+        ink_name: "Kon-peki",
+        maker: "Pilot",
+        comment: "Vivid blue, one of my faves",
+        private_comment: "Bought at such and such for X amount"
+      }
+    },
+    {
+      values: {
+        brand_name: "Pilot",
+        line_name: "Iroshizuku",
+        ink_name: "Shin-kai",
+        maker: "Pilot",
+        comment: "GOAT blue-black",
+        private_comment: "The Pen Addict made me do it"
+      }
+    },
+    {
+      values: {
+        brand_name: "Platinum",
+        line_name: undefined,
+        ink_name: "Carbon Black",
+        maker: "Platinum",
+        comment: "Perpetually inked in my Preppy",
+        private_comment:
+          "Cartridges bought at various points for about X amount per 4-pack"
+      }
+    },
+    {
+      values: {
+        brand_name: "Sailor",
+        line_name: "Manyo",
+        ink_name: "Haha",
+        maker: "Sailor",
+        comment: undefined,
+        private_comment: undefined
+      }
+    }
+  ];
+
+  it("matches and sorts by relevance", () => {
+    expect(fuzzyMatch(input, null, "sail")).toStrictEqual([
+      {
+        values: {
+          brand_name: "Sailor",
+          line_name: "Manyo",
+          ink_name: "Haha",
+          maker: "Sailor",
+          comment: undefined,
+          private_comment: undefined
+        }
+      },
+      {
+        values: {
+          brand_name: "Pilot",
+          line_name: "Iroshizuku",
+          ink_name: "Shin-kai",
+          maker: "Pilot",
+          comment: "GOAT blue-black",
+          private_comment: "The Pen Addict made me do it"
+        }
+      }
+    ]);
+
+    expect(fuzzyMatch(input, null, "many")).toStrictEqual([
+      {
+        values: {
+          brand_name: "Sailor",
+          line_name: "Manyo",
+          ink_name: "Haha",
+          maker: "Sailor",
+          comment: undefined,
+          private_comment: undefined
+        }
+      },
+      {
+        values: {
+          brand_name: "Platinum",
+          line_name: undefined,
+          ink_name: "Carbon Black",
+          maker: "Platinum",
+          comment: "Perpetually inked in my Preppy",
+          private_comment:
+            "Cartridges bought at various points for about X amount per 4-pack"
+        }
+      }
+    ]);
+
+    expect(fuzzyMatch(input, null, "carb")).toStrictEqual([
+      {
+        values: {
+          brand_name: "Platinum",
+          line_name: undefined,
+          ink_name: "Carbon Black",
+          maker: "Platinum",
+          comment: "Perpetually inked in my Preppy",
+          private_comment:
+            "Cartridges bought at various points for about X amount per 4-pack"
+        }
+      }
+    ]);
+  });
+});
