+++ app/javascript/src/collected_inks/table/index.js
@@ -1,0 +1,2 @@
+export * from "./CollectedInksTable";
+export * from "./CollectedInksTablePlaceholder";
